refactor: consolidate chat logic
Move most of the actual logic from lib.rs into ChatProcessor in chat.rs. also integrate tool call handling into ChatProcessor. Simplify the code.
